Split test for fmt
